+++ project/idsa/src/test/java/nl.tno.idsa/performance_checker/PerformanceCheckerTest.java
@@ -18,13 +18,13 @@ public class PerformanceCheckerTest {
     public void computeGraph() throws Exception {
         PerformanceChecker performance = new PerformanceChecker();
 
-        List<Double> id = Arrays.asList(5.0,4.0,3.0,2.0,1.0);
+        List<Float> id = Arrays.asList(5f,4f,3f,2f,1f);
 
         id.stream().forEach(iD -> {
             Integer numberOfElement = ThreadLocalRandom.current().nextInt(20,40);
-            List<Long> listValue = new ArrayList<>();
+            List<Integer> listValue = new ArrayList<>();
             for(int i = 0; i < numberOfElement ; i++){
-                listValue.add(ThreadLocalRandom.current().nextLong(0, 1000));
+                listValue.add(ThreadLocalRandom.current().nextInt(0, 1000));
             }
             listValue.sort((b,a) -> a.compareTo(b));
             PersonalPerformance pp = new PersonalPerformance(0);
@@ -35,9 +35,9 @@ public class PerformanceCheckerTest {
         performance.computeGraph();
 
 
-        List<Long> one = Arrays.asList(7L,6L,5L,2L,1L,1L,1L,1L,1L);
-        List<Long> two = Arrays.asList(10L,9L,8L,7L,2L,2L,2L,2L,2L,2L,2L);
-        List<Long> three = Arrays.asList(15L,7L,3L,1L,1L,1L);
+        List<Integer> one = Arrays.asList(7,6,5,2,1,1,1,1,1);
+        List<Integer> two = Arrays.asList(10,9,8,7,2,2,2,2,2,2,2);
+        List<Integer> three = Arrays.asList(15,7,3,1,1,1);
         PerformanceChecker performance1 = new PerformanceChecker();
         PersonalPerformance pp1 = new PersonalPerformance(0);
         one.stream().forEach(pp1::addValue);
